+++ app/Http/Controllers/admin/CategoryController.php
@@ -5,10 +5,13 @@ namespace App\Http\Controllers\admin;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\RequestCreateUpdateCategory;
 use App\Models\Category;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class CategoryController extends Controller
 {
+
     public function index(){
         $categories = Category::latest()->get();
 
@@ -23,6 +26,10 @@ class CategoryController extends Controller
     }
 
     public function edit($id){
+        if(Auth::user()->role !== "admin" && Auth::user()->role !== "editer"){
+            return redirect('/categories')->with(["message" => "Vous n'êtes pas autorisé a éditer une catégorie. Merci de contacter l'administrateur ou l'éditeur.", "status" => "error"]);
+        }
+
         $category = Category::findOrFail($id);
 
         return view('pages/admin/editCategory', ['category' => $category]);
@@ -36,6 +43,10 @@ class CategoryController extends Controller
     }
 
     public function destroy(Category $category){
+        if(Auth::user()->role !== "admin"){
+            return redirect('/categories')->with(["message" => "Vous n'êtes pas autorisé a supprimer une catégorie. Merci de contacter l'administrateur.", "status" => "error"]);
+        }
+
         $category->delete();
 
         return redirect('/categories')->with(['message' => 'Catégorie supprimé avec succès', 'status' => 'success']);
