+++ app/Http/Controllers/admin/ProjectController.php
@@ -7,10 +7,13 @@ use App\Http\Requests\RequestCreateProject;
 use App\Http\Requests\RequestEditProject;
 use App\Models\Client;
 use App\Models\Project;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ProjectController extends Controller
 {
+
     public function index(){
         $projects = Project::latest()->paginate(10);
 
@@ -20,11 +23,14 @@ class ProjectController extends Controller
     public function show($id){
         $project = Project::findOrFail($id);
 
-        // dd($project);
         return view('pages/admin/showProject', ['project' => $project]);
     }
 
     public function create(){
+        if(Auth::user()->role !== "admin" && Auth::user()->role !== "editer"){
+            return redirect('/projects')->with(["message" => "Vous n'êtes pas autorisé a créer un projet. Merci de contacter l'administrateur ou l'éditeur.", "status" => "error"]);
+        }
+
         $clients = Client::all();
 
         return view("pages/admin/createEditProject", ['clients' => $clients]);
@@ -39,6 +45,10 @@ class ProjectController extends Controller
     }
 
     public function edit($id){
+        if(Auth::user()->role !== "admin" && Auth::user()->role !== "editer"){
+            return redirect('/projects')->with(["message" => "Vous n'êtes pas autorisé a éditer ce projet. Merci de contacter l'administrateur ou l'éditeur.", "status" => "error"]);
+        }
+
         $project = Project::findOrFail($id);
         $clients = Client::all();
 
@@ -53,6 +63,10 @@ class ProjectController extends Controller
     }
 
     public function status(Project $project, Request $request){
+        if($request->user_id !== $project->user_id || $project->user->role !== "admin"){
+            return back()->with(['message' => "Vous n'êtes pas autorisé a éffectuer cette modification.", 'status' => "error"]);
+        }
+
         $validate = $request->validate([
             'status' => 'string|min:6'
         ]);
@@ -62,6 +76,10 @@ class ProjectController extends Controller
     }
 
     public function testimonial(Project $project, Request $request){
+        if($request->user_id !== $project->user_id || $project->user->role !== "admin"){ 
+            return back()->with(['message' => "Vous n'êtes pas autorisé a éffectuer cette modification.", 'status' => "error"]);
+        }
+
         $validate = $request->validate([
             'testimonial' => 'sometimes|string|min:10'
         ]);
@@ -71,6 +89,10 @@ class ProjectController extends Controller
     }
 
     public function destroy(Project $project){
+        if(Auth::user()->role !== "admin"){
+            return redirect('/projects')->with(["message" => "Vous n'êtes pas autorisé a supprimer un projet. Merci de contacter l'administrateur.", "status" => "error"]);
+        }
+
         $project->delete();
         return redirect('/projects')->with(['message' => 'Le projet a été supprimé avec succès', 'status' => 'success']);
     }
