+++ app/Http/Controllers/admin/TeamMemberController.php
@@ -6,11 +6,14 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\RequestCreateTeamMember;
 use App\Http\Requests\RequestUpdateTeamMember;
 use App\Models\TeamMember;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 
 class TeamMemberController extends Controller
 {
+
     public function index(){
         $team = TeamMember::latest()->paginate();
 
@@ -24,6 +27,10 @@ class TeamMemberController extends Controller
     }
 
     public function create(){
+        if(Auth::user()->role !== "admin" && Auth::user()->role !== "editer"){
+            return redirect('/team')->with(["message" => "Vous n'êtes pas autorisé a créer un membre. Merci de contacter l'administrateur ou l'éditeur.", "status" => "error"]);
+        }
+
         return view('pages/admin/teamMemberCreateEdit');
     }
 
@@ -41,6 +48,10 @@ class TeamMemberController extends Controller
     }
 
     public function edit($id){
+        if(Auth::user()->role !== "admin" && Auth::user()->role !== "editer"){
+            return redirect('/team')->with(["message" => "Vous n'êtes pas autorisé a éditer un membre. Merci de contacter l'administrateur ou l'éditeur.", "status" => "error"]);
+        }
+
         $team = TeamMember::findOrFail($id);
 
         return view('pages/admin/teamMemberCreateEdit', ['team' => $team]);
@@ -64,6 +75,10 @@ class TeamMemberController extends Controller
     }
 
     public function destroy(TeamMember $team){
+        if(Auth::user()->role !== "admin"){
+            return redirect('/team')->with(["message" => "Vous n'êtes pas autorisé a supprimer membre. Merci de contacter l'administrateur.", "status" => "error"]);
+        }
+
         $team->delete();
         return redirect('/team')->with(['message' => 'Un membre a été supprimé avec succès', 'status' => 'success']);
     }
